a bit of clean up
